Handling default networks (bridge, null, host) with local store
With the introduction of Local store, we store all the networks and
endpoints created for all the locally scoped objects.
But, the default networks are special. On daemon start, docker will
install the 3 default networks (docker0 bridge, null and host). But,
since we restore the networks from the local store in controller.New()
the actual network creation from Daemon will fail. Hence, we need to
handle that scenario. To add more to it, the user could have changed
the bridge configuration in docker daemon configuration. Hence the
NewNetwork call from the docker daemon must be honored and at the same
time we have to retain the network ID and endpoints stored in the
local-db.

The solution is to introduce a new state called `started` for the
controller & playback the network and endpoint states from the datastore
only when the Start() is called. Between the New() and Start(), the
Daemon can create the special networks which will be adjusted with the
local states & installed by merging the latest configs with the saved
data.

Signed-off-by: Madhu Venugopal <madhu@docker.com>